demo/requirements.txt
flask>=3.0.0
anthropic>=0.26.0
pillow>=10.0.0
gunicorn>=21.2.0
